removing utilities files in import_data/command and separated services into their own files
